UDRCController:  Restructure the driver
This restructures the UDRC Controller driver to ditch the awkward
HardwareInterface API.  It now directly inherits off of the
ExternalController object and provides the appropriate methods to
control the hardware.

It also ditches the extra thread in the object.  There are no lines
that we are actually providing to the rest of the program, so there's
no reason we should be polling the input pins.  If this changes in
the future, it should be some sort of interrupt driven proces.  The
output pins now just directly write.  This hopefully will not cause
too much latency in the caller's path.